Test and code on the same level again.

+++ src/maidsafe/passport/tests/system_packets_test.cc
@@ -499,7 +499,7 @@ TEST_F(SystemPacketsTest, BEH_CreateTmid) {
   uint32_t random_no_from_rid(0);
   int64_t a = 1;
   for (int i = 0; i < 4; ++i) {
-    uint8_t temp(static_cast<uint64_t>(kRid.at(i)));
+    uint8_t temp(static_cast<uint8_t>(kRid.at(i)));
     random_no_from_rid += static_cast<uint32_t>(temp * a);
     a *= 256;
   }
